refactor callbackurlschema and callback task
