+++ plugins/optimizer/objective_functions/hinge_loss/routes.py
@@ -27,9 +27,9 @@ from plugins.optimizer.coordinator.shared_schemas import (
     ObjectiveFunctionCallbackData,
     ObjectiveFunctionCallbackSchema,
 )
+from plugins.optimizer.interaction_utils.schemas import CallbackURLData, CallbackURLSchema
+from plugins.optimizer.interaction_utils.tasks import make_callback
 from qhana_plugin_runner.api.plugin_schemas import (
-    CallbackURLData,
-    CallbackURLSchema,
     DataMetadata,
     EntryPoint,
     PluginMetadata,
@@ -37,7 +37,6 @@ from qhana_plugin_runner.api.plugin_schemas import (
     PluginType,
 )
 from qhana_plugin_runner.db.models.tasks import ProcessingTask
-from qhana_plugin_runner.requests import make_callback
 
 from . import HINGELOSS_BLP, HingeLoss
 from .schemas import (
