+++ plugins/optimizer/objective_functions/ridge_loss/routes.py
@@ -30,9 +30,9 @@ from plugins.optimizer.coordinator.shared_schemas import (
     ObjectiveFunctionCallbackData,
     ObjectiveFunctionCallbackSchema,
 )
+from plugins.optimizer.interaction_utils.schemas import CallbackURLData, CallbackURLSchema
+from plugins.optimizer.interaction_utils.tasks import make_callback
 from qhana_plugin_runner.api.plugin_schemas import (
-    CallbackURLData,
-    CallbackURLSchema,
     DataMetadata,
     EntryPoint,
     PluginMetadata,
@@ -40,7 +40,6 @@ from qhana_plugin_runner.api.plugin_schemas import (
     PluginType,
 )
 from qhana_plugin_runner.db.models.tasks import ProcessingTask
-from qhana_plugin_runner.requests import make_callback
 
 from . import RIDGELOSS_BLP, RidgeLoss
 from .schemas import (
