+++ qhana_plugin_runner/api/plugin_schemas.py
@@ -409,19 +409,6 @@ class PluginMetadataSchema(MaBaseSchema):
         return PluginMetadata(**data)
 
 
-@dataclass
-class CallbackURLData:
-    callback_url: str
-
-
-class CallbackURLSchema(MaBaseSchema):
-    callback_url = ma.fields.URL(required=True, allow_none=False, data_key="callbackUrl")
-
-    @ma.post_load
-    def make_object(self, data, **kwargs):
-        return CallbackURLData(**data)
-
-
 class NumpyArray(ma.fields.Field):
     def _serialize(self, value, attr, obj, **kwargs):
         if value is None:
